+++ guillotina/api/ws.py
@@ -8,12 +8,12 @@ from guillotina import jose
 from guillotina import logger
 from guillotina.api.service import Service
 from guillotina.browser import Response
+from guillotina.interfaces import IInteraction
+from guillotina.interfaces import IPermission
 from guillotina.interfaces import ISite
 from guillotina.interfaces import ITraversableView
 from zope.component import getUtility
 from zope.component import queryMultiAdapter
-from zope.security.interfaces import IInteraction
-from zope.security.interfaces import IPermission
 
 import aiohttp
 import asyncio