+++ guillotina/browser.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from guillotina import configure
 from guillotina.interfaces import IAbsoluteURL
+from guillotina.interfaces import ILocation
 from guillotina.interfaces import IRequest
 from guillotina.interfaces import IResource
 from guillotina.interfaces import ISerializableException
@@ -8,7 +9,6 @@ from guillotina.interfaces import IView
 from guillotina.transactions import get_current_request
 from zope.component import adapter
 from zope.interface import implementer
-from zope.location import ILocation
 
 
 def get_physical_path(context):