+++ guillotina/interfaces/exceptions.py
@@ -1,7 +1,8 @@
-from zope.interface import Interface
+from zope.interface.common.interfaces import IAttributeError
+from zope.interface.common.interfaces import IException
 
 
-class ISerializableException(Interface):
+class ISerializableException(IException):
     """
     An exception that can be deserialized
     """
@@ -11,3 +12,15 @@ class ISerializableException(Interface):
         return json serializable data to be used
         with exception data in responses
         """
+
+
+class IUnauthorized(IException):
+    pass
+
+
+class IForbidden(IException):
+    pass
+
+
+class IForbiddenAttribute(IForbidden, IAttributeError):
+    pass