+++ guillotina/testing.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from aiohttp.test_utils import make_mocked_request
-from guillotina.auth.policy import Interaction
 from guillotina.auth.users import ROOT_USER_ID
 from guillotina.auth.users import RootUser
 from guillotina.browser import View
@@ -14,6 +13,7 @@ from guillotina.interfaces import IDefaultLayer
 from guillotina.interfaces import IRequest
 from guillotina.interfaces import IResource
 from guillotina.jsonfield import JSONField
+from guillotina.security.policy import Interaction
 from zope.component import getUtility
 from zope.interface import implementer
 from zope.schema import List