+++ guillotina/transactions.py
@@ -5,6 +5,7 @@ on a single ZODB connection. This should not be possible to do safely, but
 we'll see how far we get and learn more about ZODB while doing it...
 """
 from guillotina.exceptions import RequestNotFound
+from guillotina.exceptions import Unauthorized
 from guillotina.interfaces import SHARED_CONNECTION
 from guillotina.utils import get_authenticated_user_id
 from transaction._manager import _new_transaction
@@ -12,7 +13,6 @@ from transaction.interfaces import ISavepoint
 from transaction.interfaces import ISavepointDataManager
 from zope.interface import implementer
 from zope.proxy import ProxyBase
-from zope.security.interfaces import Unauthorized
 
 import asyncio
 import inspect